MAINT: Move scimath and user_array implementations
